Created Events for Brother and Sister
Created two scenes: one with the brother and one with the sister. THey
are not added into the game, but exist. They also do not display images
and also need a buffer for the spacebar too.

They should have an element on their screen/stuff happening during these
events.   Did not try to add in yet since unable to debug since no
appropriate time to run events.